requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
sentence-transformers
torchvision